More reorganization of code
