+++ lyrebird/src/main/java/moe/lyrebird/view/screens/media/handlers/EmbeddedMediaViewHelper.java
@@ -28,7 +28,7 @@ import moe.lyrebird.view.screens.Screen;
 import moe.lyrebird.view.screens.media.MediaEmbeddingService;
 import moe.lyrebird.view.screens.media.display.MediaDisplayScreen;
 import moe.lyrebird.view.screens.media.display.MediaScreenController;
-import moe.lyrebird.view.viewmodel.Clipping;
+import moe.lyrebird.view.viewmodel.javafx.Clipping;
 
 import javafx.application.Platform;
 import javafx.scene.Node;
